Option for keeping the searched value
I want to search for multiple entries and only need the select events.

The update of the input field is bad for an continuus search, since the search result would be limited to the selected item.
Therefore I added the options to clear the search after an selection or to just keep the input.

E.g.
Searching for `C` and selecting `California` can now keep `C` as search string and the user can easily add `North Carolina` without having the remove characters from the search bar. Works great with `focusAfterSelect=true`.

Im very bad at naming things, so maybe there might be better ways to do that.