Consistent id's with svelte-search
Pass Typeahead's id property to Search to ensure proper `aria-labelledby` references. 
Updated Typeahead wrapper div id to not clash with Search's input id (this feels slightly hackish to me, but avoids separate PRs to Search).

Removed an unreferenced (and oddly defined?) CSS class, updated an inner aria-labelledby reference (this is honestly a guess, as I'm not exactly an aria expert...) and ensured unique id's on the result list items.